migrate ProNewCompaniesAccesses
